pos n color gs helper

+++ src/game/position.rs
@@ -1,6 +1,6 @@
 pub enum Position{
     Bar,
-    Point(u8),
+    Point(usize),
     Out,
 }
